+++ backend/src/Controller/User/EsiController.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnused */
 
 declare(strict_types=1);
 
@@ -61,10 +62,11 @@ class EsiController extends BaseController
      *     path="/user/esi/request",
      *     operationId="request",
      *     summary="ESI request.",
-     *     description="Needs role: user-admin
-     *                  Example route: /characters/{character_id}/stats/
-     *                  Only for GET request.
-     *                  Only the {character_id} placeholder is implemented.",
+     *     description="Needs role: user-admin<br>
+     *                  Example route: /characters/{character_id}/stats/<br>
+     *                  Only for GET request.<br>
+     *                  {character_id}, {corporation_id} and {alliance_id} are automatically replaced with the
+                        corresponding IDs of the selected character",
      *     tags={"ESI"},
      *     security={{"Session"={}}},
      *     @OA\Parameter(
@@ -101,11 +103,16 @@ class EsiController extends BaseController
      *     )
      * )
      */
-    public function request(ServerRequestInterface $request): ResponseInterface
+    public function request(ServerRequestInterface $request, string $method = 'GET'): ResponseInterface
     {
         $charId = $this->getQueryParam($request, 'character', '');
         $route = $this->getQueryParam($request, 'route', '');
         $debug = $this->getQueryParam($request, 'debug') === 'true';
+
+        $body = null;
+        if ($method === 'POST') {
+            $body = $request->getBody()->__toString();
+        }
 
         // validate input
         if ($route === '' || $charId === '') {
@@ -140,8 +147,9 @@ class EsiController extends BaseController
         $token = $this->tokenService->getToken($character);
         $response = null;
         try {
-            $response = $this->httpClient->request('GET', $baseUri . $path, [
-                'headers' => ['Authorization' => 'Bearer ' . $token]
+            $response = $this->httpClient->request($method, $baseUri . $path, [
+                'headers' => ['Authorization' => 'Bearer ' . $token],
+                'body' => $body,
             ]);
         } catch (ClientException $ce) {
             return $this->prepareResponse($ce->getMessage(), $debug, $ce->getResponse(), 400);
@@ -164,6 +172,60 @@ class EsiController extends BaseController
         }
 
         return $this->prepareResponse($body, $debug, $response);
+    }
+
+    /**
+     * @OA\Post(
+     *     path="/user/esi/request",
+     *     operationId="requestPost",
+     *     summary="Same as GET /user/esi/request, but for POST requests.",
+     *     tags={"ESI"},
+     *     security={{"Session"={}}},
+     *     @OA\Parameter(
+     *         name="character",
+     *         in="query",
+     *         description="EVE character ID.",
+     *         @OA\Schema(type="integer")
+     *     ),
+     *     @OA\Parameter(
+     *         name="route",
+     *         in="query",
+     *         description="The ESI route.",
+     *         @OA\Schema(type="string")
+     *     ),
+     *     @OA\Parameter(
+     *         name="debug",
+     *         in="query",
+     *         description="Show all headers, do not use cache",
+     *         @OA\Schema(type="string", enum={"true", "false"})
+     *     ),
+     *     @OA\RequestBody(
+     *         required=true,
+     *         description="JSON encoded data.",
+     *         @OA\MediaType(
+     *             mediaType="text/plain",
+     *             @OA\Schema(type="string")
+     *         ),
+     *     ),
+     *     @OA\Response(
+     *         response="200",
+     *         description="The result from ESI or an error message.",
+     *         @OA\JsonContent(type="string")
+     *     ),
+     *     @OA\Response(
+     *         response="400",
+     *         description="Error.",
+     *         @OA\JsonContent(type="string")
+     *     ),
+     *     @OA\Response(
+     *         response="403",
+     *         description="Not authorized."
+     *     )
+     * )
+     */
+    public function requestPost(ServerRequestInterface $request): ResponseInterface
+    {
+        return $this->request($request, 'POST');
     }
 
     /**